UCX/PML/SPML/OSC: fix compiler warnings
(cherry picked from commit 334c6ec3a04a6f9f2ba3fea4d8c6d47c956b94ee)

+++ oshmem/mca/spml/ucx/spml_ucx.h
@@ -244,9 +244,9 @@ extern size_t mca_spml_ucx_test_some(void *ivars, int cmp, void *cmp_value,
         size_t nelems, size_t *indices, const int *status, int datatype);
 extern int mca_spml_ucx_test_all_vector(void *ivars, int cmp, void
         *cmp_values, size_t nelems, const int *status, int datatype);
-extern size_t mca_spml_ucx_test_any_vector(void *ivars, int cmp, void
+extern int mca_spml_ucx_test_any_vector(void *ivars, int cmp, void
         *cmp_values, size_t nelems, const int *status, int datatype);
-extern size_t mca_spml_ucx_test_some_vector(void *ivars, int cmp, void
+extern int mca_spml_ucx_test_some_vector(void *ivars, int cmp, void
         *cmp_values, size_t nelems, size_t *indices, const int *status, int
         datatype);
 extern int mca_spml_ucx_team_sync(shmem_team_t team);
